requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.2
langchain>=0.0.350
openai>=1.3.5
anthropic>=0.7.4
google-cloud-aiplatform>=1.36.4
chromadb>=0.4.18
snowflake-connector-python>=3.6.0
google-cloud-bigquery>=3.13.0
databricks-sql-connector>=3.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
prometheus-client>=0.19.0
redis>=5.0.1
sqlalchemy>=2.0.23
alembic>=1.13.0
pytest>=7.4.3
jupyter>=1.0.0
pandas>=2.1.3
numpy>=1.26.2 